+ вынес редактор в отдельный компонент
