+++ packages/client/src/components/organisms/forum/forum-editor/index.ts
@@ -1,0 +1,1 @@
+export * from './forum-editor.component'
