+++ packages/client/src/components/organisms/forum/forum-post.components.tsx
@@ -1,8 +1,5 @@
 import { ForumPostProps } from './forum-types'
 import {
-  FormButtonWrapper,
-  FormSeparator,
-  FormTextarea,
   ForumPostBlock,
   ForumPostBlockAvatar,
   ForumPostBottom,
@@ -18,13 +15,10 @@ import {
 } from '@/components/templates/forum/forum.styles'
 import IconRateMinus from '@/assets/icons/rate_minus.svg'
 import IconRatePlus from '@/assets/icons/rate_plus.svg'
-import { ReactElement, useRef, useState } from 'react'
-import useSectionForm from './forum-logic'
-import { Button } from '@/components'
+import { useState } from 'react'
 import dateParse from '@/utils/dateParse'
-import { H2 } from '@/global-styles'
 import avatarDefault from '@/assets/images/avatarDefault.png'
-import { useParams } from 'react-router-dom'
+import { ForumEditor } from '@/components'
 
 const ForumPost = (el: ForumPostProps) => {
   const [rate, setRate] = useState(el.rate)
@@ -60,7 +54,8 @@ const ForumPost = (el: ForumPostProps) => {
           </ForumPostRate>
         </ForumPostBottom>
         {replyOpen && (
-          <ForumPostsForm
+          <ForumEditor
+            title="Ответить на сообщение"
             replyMessage={
               <>
                 <strong>{userName}</strong>
@@ -82,55 +77,5 @@ export const ForumPosts = ({ data }: { data: ForumPostProps[] }) => {
         <ForumPost key={el.id} {...el} />
       ))}
     </div>
-  )
-}
-
-export const ForumPostsForm = ({
-  replyMessage,
-}: {
-  replyMessage?: ReactElement
-}) => {
-  const { id, postPageId } = useParams()
-  const { register, onSubmitHandler, handleSubmit, isValid, setValue } =
-    useSectionForm(id!, postPageId)
-  const messageRef = useRef(null)
-  const messageClear = () =>
-    ((messageRef!.current! as HTMLDivElement).innerHTML = '')
-
-  return (
-    <form onSubmit={handleSubmit(onSubmitHandler)}>
-      <FormSeparator />
-      <H2 marginBottom="30px">Написать сообщение</H2>
-      <FormTextarea
-        {...register('message', { required: true })}
-        suppressContentEditableWarning={true}
-        onInput={e => {
-          setValue('message', e.currentTarget.innerHTML, {
-            shouldValidate: true,
-          })
-        }}
-        contentEditable="true"
-        ref={messageRef}
-        placeholder="Сообщение...">
-        {replyMessage && (
-          <>
-            <blockquote>{replyMessage}</blockquote>
-            <br />
-            <br />
-          </>
-        )}
-      </FormTextarea>
-      <FormButtonWrapper>
-        <Button
-          as="button"
-          type="submit"
-          color="#579945"
-          variant="contained"
-          onClick={() => messageClear()}
-          disabled={!isValid}>
-          Отправить
-        </Button>
-      </FormButtonWrapper>
-    </form>
   )
 }
