+++ packages/client/src/components/organisms/forum/index.ts
@@ -1,3 +1,4 @@
 export * from './forum-start.components'
 export * from './forum-section.components'
 export * from './forum-post.components'
+export * from './forum-editor'
